requirements.txt
discord.py==2.3.2
Flask==2.2.2